rename xray to xraysim
