+++ tex-ray/main.py
@@ -1,5 +1,5 @@
 import tifffile
-from x_ray import generate_sinograms
+from x_ray_simulation import generate_sinograms
 from tomographic_reconstruction import perform_tomographic_reconstruction
 from textile_generation import generate_unit_cell
 from zeiss_xradia_410_versa import read_txrm_scan_data, read_txm_scan_data
